requirements.txt
datasets
evaluate
matplotlib
openpyxl
pandas
scikit-learn
sentencepiece
setproctitle
simple-parsing
torch-optimizer
torchaudio
torchvision
tqdm
transformers
wandb
